pyproject.toml
[project]
name = "finance-tracker-be"
version = "0.1.0"
description = "Backend application for the finance tracker project"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.1.7",
    "djangorestframework>=3.16.0",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "django-silk>=5.3.2",
    "drf-spectacular>=0.28.0",
    "nplusone>=1.0.0",
    "pre-commit>=4.2.0",
]
